update form customer detail in cart page

+++ js/cart.js
@@ -2,6 +2,8 @@ let carts = JSON.parse(localStorage.getItem('cart')) || [];
 const fullname = document.getElementById("fullname");
 const email = document.getElementById("email");
 const address = document.getElementById("address");
+const district = document.getElementById("district");
+const provice = document.getElementById("province");
 const phone = document.getElementById("phone");
 
 if (carts.length == 0) {
@@ -108,6 +110,10 @@ function validatePhone(phone) {
     return phonePattern.test(phone); // Check if phone number matches the pattern
 }
 
+function validateDistrict(district) {
+    return district.trim().length > 0; // Check if district is not empty
+}
+
 // Address Validation
 function validateAddress(address) {
     return address.trim().length > 0; // Check if address is not empty
@@ -117,9 +123,9 @@ function setFullName(textName){
     const isFullNameValid = validateFullName(textName);
         
     if (!isFullNameValid) {
-        document.getElementById("notifyName").innerHTML = `The username must be more than one word.`;
+        document.getElementById("notifyName").innerHTML = "The username must be more than one word.";
     }else{
-        document.getElementById("notifyName").innerHTML = ``;
+        document.getElementById("notifyName").innerHTML = "";
     }
 }
 
@@ -127,9 +133,9 @@ function setEmail(textEmail){
     const isEmailValid = validateEmail(textEmail);
         
     if (!isEmailValid) {
-        document.getElementById("notifyEmail").innerHTML = `The email address is invalid`;
+        document.getElementById("notifyEmail").innerHTML = "The email address is invalid";
     }else{
-        document.getElementById("notifyEmail").innerHTML = ``;
+        document.getElementById("notifyEmail").innerHTML = "";
     }
 }
 
@@ -137,9 +143,9 @@ function setPhone(textPhone){
     const isPhoneValid = validatePhone(textPhone);
         
     if (!isPhoneValid) {
-        document.getElementById("notifyPhone").innerHTML = `Phone number must be 10 digits.`;
+        document.getElementById("notifyPhone").innerHTML = "Phone number must be 10 digits.";
     }else{
-        document.getElementById("notifyPhone").innerHTML = ``;
+        document.getElementById("notifyPhone").innerHTML = "";
     }
 }
 
@@ -147,9 +153,27 @@ function setAddress(textAddress) {
     const isAddressValid = validateAddress(textAddress);
 
     if (!isAddressValid) {
-        document.getElementById("notifyAddress").innerHTML = `The address is invalid`;
+        document.getElementById("notifyAddress").innerHTML = "The address is invalid";
     }else{
-        document.getElementById("notifyAddress").innerHTML = ``;
+        document.getElementById("notifyAddress").innerHTML = "";
+    }
+}
+
+function setProvince(textProvince) {
+    if (textProvince === "none") {
+        document.getElementById("notifyProvince").innerHTML = "The province is required";
+    } else {
+        document.getElementById("notifyProvince").innerHTML = "";
+    }
+}
+
+function setDistrict(textDistrict){
+    const isDistrictValid = validateDistrict(textDistrict);
+
+    if (!isDistrictValid) {
+        document.getElementById("notifyDistrict").innerHTML = "The district is required";
+    }else{
+        document.getElementById("notifyDistrict").innerHTML = "";
     }
 }
 
@@ -158,20 +182,24 @@ function onCheckout() {
         name: fullname.value,
         email: email.value,
         address: address.value,
+        district: district.value,
+        province: provice.value,
         phone: phone.value
     };
-    
+
     const isFullNameValid = validateFullName(user.name);
     const isEmailValid = validateEmail(user.email);
     const isPhoneValid = validatePhone(user.phone);
     const isAddressValid = validateAddress(user.address);
+    let isDistrict = user.district;
+    let isExistProvince = user.province;
     
     if (carts.length != 0) {
-        if (isFullNameValid && isEmailValid && isPhoneValid && isAddressValid) {
+        if (isFullNameValid && isEmailValid && isPhoneValid && isAddressValid && isDistrict != "" && isExistProvince != "none") {
             localStorage.setItem("username", JSON.stringify(user));
             window.location.href = "checkout.html"; // Redirect to checkout
         }else {
-            if(!isFullNameValid && !isEmailValid && !isPhoneValid && !isAddressValid){
+            if(!isFullNameValid && !isEmailValid && !isPhoneValid && !isAddressValid && !isDistrict && isExistProvince === "none"){
                 alert("All fields are required!");
             }
         }
